added user_index_button for admins and users#index action

+++ app/helpers/application_helper.rb
@@ -20,7 +20,7 @@ module ApplicationHelper
   end
 
   def delete_button(item)
-    if !!current_user.admin
+    if !!current_user && current_user.admin
       if item.class != User
         link_to "Delete #{item.class}", item, method: :delete, data: { confirm: "Are you sure you want to delete #{item.full_title}?" }
       else
